modify field related functions to methods of messageBuilder
