+++ adapter/protobuf/message.go
@@ -16,98 +16,6 @@ type message struct {
 	isCycled bool
 }
 
-type messageBuilder struct {
-	m *message
-	d *desc.MessageDescriptor
-
-	dependOn map[string]bool
-
-	// used to detect cycle fields
-	usedMessage map[string]entity.Message
-}
-
-func (b *messageBuilder) processMessageField(f *desc.FieldDescriptor) {
-	field := &messageField{
-		d: f,
-	}
-
-	// self-referenced
-	if f.GetMessageType().GetName() == b.m.Name() {
-		b.m.isCycled = true
-		field.Message = b.m
-		b.add(field)
-		return
-	}
-
-	if m, ok := b.usedMessage[f.GetMessageType().GetName()]; ok {
-		b.m.isCycled = true
-		field.Message = m
-		b.add(field)
-		return
-	}
-
-	b2 := &messageBuilder{
-		m: &message{
-			d: f.GetMessageType(),
-		},
-		d:           f.GetMessageType(),
-		usedMessage: b.usedMessage,
-	}
-
-	field.Message = b2.build()
-
-	b.usedMessage[f.GetMessageType().GetName()] = field.Message
-	b.add(field)
-}
-
-func (b *messageBuilder) add(f entity.Field) {
-	b.m.fields = append(b.m.fields, f)
-}
-
-func (b *messageBuilder) build() entity.Message {
-	// collect messages and enums which declared in the target message
-
-	msgs := make([]entity.Message, 0, len(b.d.GetNestedMessageTypes()))
-	for _, d := range b.d.GetNestedMessageTypes() {
-		msgs = append(msgs, newMessage(d))
-	}
-	b.m.nestedMessages = msgs
-
-	enums := make([]entity.Enum, 0, len(b.d.GetNestedEnumTypes()))
-	for _, d := range b.d.GetNestedEnumTypes() {
-		enums = append(enums, newEnum(d))
-	}
-	b.m.nestedEnums = enums
-
-	// it need to resolve oneofs before resolve other fields.
-	// GetFields contains fields of oneofs.
-	encounteredOneOfFields := map[string]bool{}
-	for _, o := range b.d.GetOneOfs() {
-		for _, c := range o.GetChoices() {
-			encounteredOneOfFields[c.GetFullyQualifiedName()] = true
-		}
-		b.add(newOneOfField(o))
-	}
-
-	// TODO: label, options
-	for _, f := range b.d.GetFields() {
-		// skip fields of oneofs
-		if encounteredOneOfFields[f.GetFullyQualifiedName()] {
-			continue
-		}
-
-		// self
-
-		if isMessageType(f.GetType()) {
-			b.processMessageField(f)
-		} else {
-			b.add(newField(f))
-		}
-	}
-
-	return b.m
-}
-
 func newMessage(d *desc.MessageDescriptor) entity.Message {
 	msg := &message{
 		d:      d,
